mandatory part

+++ client.c
@@ -6,7 +6,7 @@
 /*   By: jtoumani <jtoumani@student.42heilbronn.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/11 12:02:55 by jtoumani          #+#    #+#             */
-/*   Updated: 2025/06/16 15:58:58 by jtoumani         ###   ########.fr       */
+/*   Updated: 2025/06/18 18:03:00 by jtoumani         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,20 +23,21 @@ void	send_signal(int pid, char *str)
 		bit = 0;
 		while (bit < 8)
 		{
-			if ((str[i] >> bit) & 1)
-				kill(pid, SIGUSR2);
-			else
+			if ((str[i] >> (7 - bit)) & 1)
 				kill(pid, SIGUSR1);
+			else
+				kill(pid, SIGUSR2);
 			usleep(100);
 			bit++;
 		}
 		i++;
 	}
 	bit = 0;
-	while (bit++ < 8)
+	while (bit < 8)
 	{
-		kill(pid, SIGUSR1);
+		kill(pid, SIGUSR2);
 		usleep(100);
+		bit++;
 	}
 }
 int	main(int argc, char **argv)
@@ -46,6 +47,8 @@ int	main(int argc, char **argv)
 	if (argc == 3)
 	{
 		pid = ft_atoi(argv[1]);
+		if(pid <= 0)
+			exit(1);
 		str = argv[2];
 		send_signal(pid, str);
 	}
